Build Index vue for views
Vue starter kit is insalled.  Use Inertia.js for Vue components served
via Laravel controllers.  Inertia.js glues Laravel and Vue together.

Vue ← Inertia ← Laravel Controller (returns Inertia::render())

Manually create the .vue files directly for Index.vue component.

Use defineProps() which is a compile-time macro in Vue <script setup>
that declares which props the component accepts from its parent.  This
is how a Vue component receives data from outside for it to accept.

This means the Index.vue component expects to receive a prop by the name
of posts and it should be of type Object.  Actually, in this case, posts
is a Laravel pagination object — which is a special kind of object with
.data, .links, .meta, etc.

Use named import of Link to build and use <InertiaLink>...</InertiaLink>
for pagination.