Use TypeScript for posts props
We should use TypeScript in lieu of JavaScript to declare defineProps
for type safety without mixing compile-time and runtime type checking.

In TypeScript the posts type is required by default, making it obsolete
for required: true.

The interface name doesn't matter.  The important part is the structure
inside the interface.  The generic <InterfaceName> just tells TypeScript
which interface to use.  Using PostsType as the interface name is an
excellent choice.