Use route() helper for sidebar links
In Laravel, the global route() helper function generates a URL based on
a named route from the Laravel back end.  This writes the URL directly
in the front end, and we can reference the name of the route defined in
routes/web.php.  If we change the URL later (e.g., to /articles), only
update it in routes/web.php and the front end updates automatically.