Build Welcome link in sidebar
We use the name of 'posts.index' as the entry point to this Posts app.

Add Welcome link in the sidebar and update the '/' route with redirect
to Posts index page if the user is authenticated.

Import Globe icon for Welcome link and use a dynamic link in lieu of a
hard-coded string.

Update the link of the Laravel icon and point it to Posts index page in
alignment with the new sidebar layout as the user is authenticated.  For
guest user, this link is not visible.