Use Auth::check() in lieu of global auth() helper
Intelephense doesn't know that auth() returns an instance of AuthManager
or Illuminate\Contracts\Auth\Factory, which has the check() method.

In Laravel 12, Taylor Otwell made global helpers opt-in to reduce global
namespace pollution, improve performance (slightly) and encourage use of
dependency injection or facades.

Thus, use Facades in lieu of helpers.  This works out-of-the-box with no
need for helpers.php and is fully recognized by Intelephense.