Use ?? null for predictable type
The type of page.props.flash?.success might be string ("Post created
successfully.") or undefined (if no success message exists).  So the
type of the computed is ComputedRef<string | undefined>.

With ?? null, if page.props.flash?.success is undefined again, it will
be replaced with null.  Now the computed is ComputedRef<string | null>
and it will never be undefined.

Thus, ?? null normalizes optional values into a consistent nullable type
and that prevents type errors.