Extend global PageProps type with flash
We use usePage as a generic composable that provides the current Inertia
page props.

By default, Inertia doesn't know the props structure → TypeScript infers
it as {}.  In Posts Index vue component, TS complains about this:

Property 'success' does not exist on type '{}'. (tsserver 2339)

Because as far as TypeScript knows, props is {}.

PageProps is augmented globally in globals.d.ts, extending AppPageProps.
AppPageProps is defined in index.d.ts with no flash definition.  To let
TS be satisfied, we need to add flash to the existing AppPageProps type.