Define the delete rule in PostPolicy
The $post->user_id is idiomatic as the posts table has a user_id foreign
key.  When we fetch a Post, Laravel always loads user_id as it's part of
the model's attributes.

Use $post->user_id for policy rule as it is always available, zero cost,
and the standard way of policy ownership checks.