Add flash.info support to Index Vue Component
Currently we only watch and display flash.success.  Update localFlash to
also handle flash.info as well.

Update template to show both types.  Now flash will display both success
and info messages.

By default, watch() is lazy and only runs the callback when the watched
value changes, not when the component first mounts.

By adding { immediate: true }, it forces the watcher to run immediately
and to use the current (initial) value when the component is created.
We need { immediate: true } for flash messages because Inertia may have
flash data when the page loads!